requirements.txt
# Core video processing
opencv-python>=4.5.0
numpy>=1.21.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Audio transcription
openai-whisper>=20231117
torch>=1.9.0
torchaudio>=0.9.0

# Video downloading
pytube>=15.0.0
requests>=2.25.0

# Image processing and analysis
scikit-image>=0.18.0
Pillow>=8.0.0

# Document generation
python-docx>=0.8.11
reportlab>=3.6.0

# Web automation (for private platforms)
selenium>=4.0.0
webdriver-manager>=3.8.0

# Utilities
tqdm>=4.60.0
pathlib2>=2.3.0

# GUI (usually comes with Python)
# tkinter - included with Python standard library

# Optional: For better performance
# torch-audio  # Uncomment if you want GPU acceleration for Whisper
